refactor(core): extract extendMaskReplacers.js to a separate file
